Tie dnanexus setuptools to 58
Newer versions of python/setuptools no longer support `use_2to3`, which is needed by pyvcf.
For now, the nexus app will just revert to a previous version of setuptools. However, what's actually needed is a
removal of pyvcf from the code as it provides nothing that can't be replaced with pysam.